The protocol is not limited anymore to a user defined maximum path
Now the server and the client make use of the string buffer dynamic
length to avoid the need of a fixed maximum path length for stack
buffers.

This brings the maximum path length supported to the maximum OS path
length (under windows the 256 limit is removed in the files_win32.c
module).